Rename 'FemLinearSystem2' to 'DoFLinearSystem' and 'FemLinearSystem' to 'NodeLinearSystem'.
