+++ femutils/NodeLinearSystem.h
@@ -5,12 +5,12 @@
 // SPDX-License-Identifier: Apache-2.0
 //-----------------------------------------------------------------------------
 /*---------------------------------------------------------------------------*/
-/* FemLinearSystem2.h                                          (C) 2022-2022 */
+/* NodeLinearSystem.h                                           (C) 2022-2022 */
 /*                                                                           */
-/* Linear system: Matrix A + Vector x + Vector b for Ax=b wit DoFs.          */
+/* Linear system: Matrix A + Vector x + Vector b for Ax=b.                   */
 /*---------------------------------------------------------------------------*/
-#ifndef FEMTEST_FEMLINEARSYSTEM2_H
-#define FEMTEST_FEMLINEARSYSTEM2_H
+#ifndef FEMTEST_FEMLINEARSYSTEM_H
+#define FEMTEST_FEMLINEARSYSTEM_H
 /*---------------------------------------------------------------------------*/
 /*---------------------------------------------------------------------------*/
 
@@ -21,21 +21,19 @@
 /*---------------------------------------------------------------------------*/
 /*!
  * \internal.
- * \brief Implementation for FemLinearSystem.
+ * \brief Implementation for NodeLinearSystem.
  */
-class FemLinearSystem2Impl
+class NodeLinearSystemImpl
 {
  public:
 
-  virtual ~FemLinearSystem2Impl() = default;
+  virtual ~NodeLinearSystemImpl() = default;
 
  public:
 
-  virtual void matrixAddValue(Arcane::DoFLocalId row, Arcane::DoFLocalId column, Arcane::Real value) = 0;
+  virtual void matrixAddValue(Arcane::NodeLocalId row, Arcane::NodeLocalId column, Arcane::Real value) = 0;
   virtual void setRHSValues(Arcane::Span<const Arcane::Real> values) = 0;
   virtual void solve() = 0;
-  virtual Arcane::VariableDoFReal& solutionVariable() = 0;
-  virtual Arcane::VariableDoFReal& rhsVariable() = 0;
 };
 
 /*---------------------------------------------------------------------------*/
@@ -47,34 +45,30 @@ class FemLinearSystem2Impl
  * initialize(). If you want to reuse the same instance for several solving
  * you need to call reset() to destroy the underlying linear system and then
  * you need to call initialize() again.
- *
- * The solve() method solves the current linear system. After this variable
- * returned by the method solutionVariable() will be filled with the values
- * of the solution vector.
  */
-class FemLinearSystem2
+class NodeLinearSystem
 {
  public:
 
-  FemLinearSystem2();
-  ~FemLinearSystem2();
-  FemLinearSystem2(const FemLinearSystem2&) = delete;
-  FemLinearSystem2(FemLinearSystem2&&) = delete;
-  FemLinearSystem2& operator=(FemLinearSystem2&&) = delete;
-  FemLinearSystem2& operator=(const FemLinearSystem2&) = delete;
+  NodeLinearSystem();
+  ~NodeLinearSystem();
+  NodeLinearSystem(const NodeLinearSystem&) = delete;
+  NodeLinearSystem(NodeLinearSystem&&) = delete;
+  NodeLinearSystem& operator=(NodeLinearSystem&&) = delete;
+  NodeLinearSystem& operator=(const NodeLinearSystem&) = delete;
 
  public:
 
   /*!
    * \brief Initialize the instance.
    *
-   * The variable dof_variable will be filled with the solution value after
+   * The variable node_variable will be filled with the solution value after
    * the call to the method solve().
    */
-  void initialize(Arcane::ISubDomain* sd, Arcane::IItemFamily* dof_family, const Arcane::String& solver_name);
+  void initialize(Arcane::ISubDomain* sd, const Arcane::VariableNodeReal& node_variable);
 
   //! Add the value \a value to the (row,column) element of the matrix
-  void matrixAddValue(Arcane::DoFLocalId row, Arcane::DoFLocalId column, Arcane::Real value);
+  void matrixAddValue(Arcane::NodeLocalId row, Arcane::NodeLocalId column, Arcane::Real value);
 
   /*!
    * \brief Set the values for vector B.
@@ -95,25 +89,9 @@ class FemLinearSystem2
    */
   void reset();
 
-  /*!
-   * \brief Variable containing the solution vector.
-   *
-   * The values of this variable are only relevant after a call to solve().
-   */
-  Arcane::VariableDoFReal& solutionVariable();
-
-  /*!
-   * \brief Variable containing the right hand side vector.
-   *
-   * The values of this variable will be used during the solve() call to
-   * fill the right hand side vector.
-   */
-  Arcane::VariableDoFReal& rhsVariable();
-
  private:
 
-  FemLinearSystem2Impl* m_p = nullptr;
-  Arcane::IItemFamily* m_item_family = nullptr;
+  NodeLinearSystemImpl* m_p = nullptr;
 
  private:
